+++ internal/service/workspacesweb/exports_test.go
@@ -5,11 +5,15 @@ package workspacesweb
 
 // Exports for use in tests only.
 var (
-	ResourceBrowserSettings = newBrowserSettingsResource
-	ResourceNetworkSettings = newNetworkSettingsResource
-	ResourceUserSettings    = newUserSettingsResource
+	ResourceBrowserSettings        = newBrowserSettingsResource
+	ResourceNetworkSettings        = newNetworkSettingsResource
+	ResourceUserSettings           = newUserSettingsResource
+	ResourceDataProtectionSettings = newDataProtectionSettingsResource
+	ResourceIPAccessSettings       = newIPAccessSettingsResource
 
-	FindBrowserSettingsByARN = findBrowserSettingsByARN
-	FindNetworkSettingsByARN = findNetworkSettingsByARN
-	FindUserSettingsByARN    = findUserSettingsByARN
+	FindBrowserSettingsByARN        = findBrowserSettingsByARN
+	FindNetworkSettingsByARN        = findNetworkSettingsByARN
+	FindUserSettingsByARN           = findUserSettingsByARN
+	FindDataProtectionSettingsByARN = findDataProtectionSettingsByARN
+	FindIPAccessSettingsByARN       = findIPAccessSettingsByARN
 )