+++ internal/service/workspacesweb/service_package_gen.go
@@ -29,6 +29,22 @@ func (p *servicePackage) FrameworkResources(ctx context.Context) []*types.Servic
 			},
 		},
 		{
+			Factory:  newDataProtectionSettingsResource,
+			TypeName: "aws_workspacesweb_data_protection_settings",
+			Name:     "Data Protection Settings",
+			Tags: &types.ServicePackageResourceTags{
+				IdentifierAttribute: "data_protection_settings_arn",
+			},
+		},
+		{
+			Factory:  newIPAccessSettingsResource,
+			TypeName: "aws_workspacesweb_ip_access_settings",
+			Name:     "IP Access Settings",
+			Tags: &types.ServicePackageResourceTags{
+				IdentifierAttribute: "ip_access_settings_arn",
+			},
+		},
+		{
 			Factory:  newNetworkSettingsResource,
 			TypeName: "aws_workspacesweb_network_settings",
 			Name:     "Network Settings",