+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## 5.100.0 (Unreleased)
+
+FEATURES:
+
+* **New Resource:** `aws_workspacesweb_data_protection_settings` ([#42852](https://github.com/hashicorp/terraform-provider-aws/issues/42852))
+* **New Resource:** `aws_workspacesweb_ip_access_settings` ([#42863](https://github.com/hashicorp/terraform-provider-aws/issues/42863))
+
 ## 5.99.1 (May 30, 2025)
 
 BUG FIXES: